moved searchQueryParamName and useQueryParameter to src
